Cleaned up the code
I think that this is at a workable point now. In the future once I start
making more and more little libraries like this I will come back and
substitute them in.

In the future as well I should probably turn it into a single header
file, it will help with integration into the other little libraries that
I plan on making.